#4 Simplify version identifier.
Up until now, the version identifier was formatted as:

    <major>.<minor>.<patch> RELEASE <increment>

where everything before 'release' matches the version of the upstream
Converse dependency. This format proves difficult to work with, as it
doesn't play nice with the combination of IgniteRealtime website, Maven,
and Openfire-based version identifier parsing.

This commit reverts back to a simpler format. We'll have to figure out
a different way to denote multiple builds based on the same version of
Converse (but to date, that hasn't happened yet - jinxing it as I write
this).